The diff output shows changes made to the Main functional component. Specifically, the changes are related to adding or removing classes to the document.body element based on the current route.
Line 6: A router constant is defined using the useRouter hook from the next/router package.
Lines 8-16: An useEffect hook is used to add or remove classes to the document.body element based on the current route. Previously, the code only added the main-bg class if the current route was the main route ("/"). Now, the main-bg class is added to the body if the current route is either "/", "/about", or "/blog". If the current route does not match any of these routes, the main-bg class is removed from the body.